guards and role base access is implemented

+++ Frontend/src/app/features/dashboard/dashboard.component.ts
@@ -4,6 +4,7 @@ import { DepartmentApiService } from '@app/core/services/department-api.service'
 import { PositionApiService } from '@app/core/services/position-api.service';
 import { RoleApiService } from '@app/core/services/role-api.service';
 import { StorageService } from '@app/core/services/storage.service';
+import { AuthService } from '@app/core/services/auth.service';
 
 @Component({
   standalone: true,
@@ -18,6 +19,9 @@ export class DashboardComponent {
   public readonly positionService = inject(PositionApiService);
   public readonly rolesService = inject(RoleApiService);
   public readonly storeService = inject(StorageService);
+  public readonly AuthService = inject(AuthService);
+
+  public readonly currentUser = this.AuthService.currentUser;
 
   public ngOnInit(): void {
     this.loadLocalData();
@@ -25,10 +29,17 @@ export class DashboardComponent {
 
   public loadLocalData(): void {
 
+    if (this.currentUser()?.role === 'Employee') {
+      return;
+    }
+    if (this.currentUser()?.role === 'Manager') {
+      return;
+    }
     this.storeService.removeItem('branches');
     this.storeService.removeItem('departments');
     this.storeService.removeItem('positions');
     this.storeService.removeItem('roles');
+
 
     this.branchService.getAllBranches().subscribe({
       next: (response) => {
